feat(logs): add logs screen with filtering and code simplification
- Add new Logs section in sidebar to view application event history
- Support filtering by level (Error/Warning/Success/Info)
- Store last 1000 log entries in circular buffer
- Toast messages now also written to persistent log
- Unify LogLevel and ToastLevel into single Level enum
- Remove unused LogEntry.id field
- Move icon/label presentation logic to UI layer

+++ src/kafka/client.rs
@@ -600,7 +600,7 @@ impl KafkaClient {
             rdsys::rd_kafka_DeleteRecords_destroy(del_records);
             rdsys::rd_kafka_AdminOptions_destroy(opts);
 
-            let event = rdsys::rd_kafka_queue_poll(queue, req_timeout_ms as i32);
+            let event = rdsys::rd_kafka_queue_poll(queue, req_timeout_ms);
             if event.is_null() {
                 rdsys::rd_kafka_queue_destroy(queue);
                 return Err(AppError::Kafka("DeleteRecords timed out".into()));
